Refactor: extract reusable ClimbsList and climb actions hook, delete PlaylistClimbsList
Decouple ClimbsList from QueueContext by accepting data via props instead of
reading from the hook directly. Extract favorites/playlists provider data
fetching from QueueContext into a reusable useClimbActionsData hook. This
allows the playlist page to reuse ClimbsList with its own data source instead
of duplicating ~580 lines of provider setup in PlaylistClimbsList.

- Create use-climb-actions-data.tsx hook for favorites + playlists providers
- Create queue-climbs-list.tsx wrapper bridging QueueContext to ClimbsList props
- Refactor ClimbsList to accept climbs, callbacks, and header via props
- Update playlist-detail-content.tsx to use ClimbsList + provider hook
- Delete playlist-climbs-list.tsx (586 lines of duplicated code)
